Collision detection kinda works
Added a collision callback to state which is send the colliding
object along with the type and direction of the collision relative
to the state the callback is registered to. Currently this is only
used to check tile collisions against the player, but the same
approach should work for monster and flag collisions, as well as
for bullet collisions against monsters.