Added bullets which kill monsters when colliding
See bullet.py. Bullets are stored inside the player's entity list,
and removed on impact with monsters. Currently an AABB is used to
check for collisions even though the bullet is round, but this isn't
a problem due to its small size. This does cause a memory leak, however,
since bullets are not currently despawned when they leave the screen.
Upon collision the monster is respawned at its initial spawn position.
This is fine as a temporary solution, but ideally the flag should instead
be spawned when the number of monsters reaches 0.